Implement Bucket List view (REX-47)
- Extract Bucket Types to component
- Refactor bucket type display
- Add bucket list template
- Add bucket list controller and route, add button

Display bucket list and totals
