+++ priv/ember_riak_explorer/app/components/riak-buckets.js
@@ -1,0 +1,4 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+});
