+++ priv/ember_riak_explorer/app/controllers/bucket-list.js
