--- priv/ember_riak_explorer/app/controllers/bucket-types.js
@@ -1,6 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Controller.extend({
-    queryParams: ['cluster_id'],
-    cluster_id: null
-});
