+++ priv/ember_riak_explorer/app/router.js
@@ -7,8 +7,8 @@ var Router = Ember.Router.extend({
 
 export default Router.map(function() {
   this.route('explorer_api');
-  this.route('bucket_types');
   this.route('bucket_type');
+  this.route('bucket_list');
   this.route('riak_ping');
   this.route('node_stats');
   this.route('cluster');
