+++ priv/ember_riak_explorer/app/routes/cluster.js
@@ -8,7 +8,11 @@ export default Ember.Route.extend({
     },
 
     model: function(params) {
-        console.log('cid: ' +params.cluster_id);
-        return this.explorer.getCluster(params.cluster_id, true);
+        var cluster = {
+            cluster: this.explorer.getCluster(params.cluster_id, true),
+            bucket_types: this.store.find('bucket_type', { cluster_id: params.cluster_id })
+        };
+
+        return new Ember.RSVP.hash(cluster);
     }
 });
