+++ priv/ember_riak_explorer/app/templates/components/bucket-types.hbs
@@ -1,0 +1,34 @@
+<table class="table">
+<tbody>
+{{#each bt in bucket_types}}
+    <tr>
+        <td>
+        <div class="container">
+            <div class="row">
+                <div class="col-md-4">{{bt.id}}</div>
+                <div class="col-md-3">
+                    <strong>Active:</strong> {{bt.props.active}},
+                    <strong>allow_mult:</strong> {{bt.props.allow_mult}}
+                </div>
+                <div class="col-md-5">
+                    <strong>Actions:</strong>
+                    {{#link-to 'bucket_type' (query-params cluster_id=cluster_id bucket_type_id=bt.id)
+                            classNames='btn btn-xs btn-primary' }}
+                        View Properties{{/link-to}}
+                    {{#link-to 'bucket_list' (query-params cluster_id=cluster_id bucket_type_id=bt.id)
+                            classNames='btn btn-xs btn-danger' }}
+                        Bucket List{{/link-to}}
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-md-12">
+                </div>
+            </div>
+        </div>
+        </td>
+    </tr>
+{{else}}
+    No bucket types. Have you specified a node?
+{{/each}}
+</tbody>
+</table>
